pyproject.toml
[project]
name = "zmk-buddy"
version = "0.1.1"
description = "A live keymap visualization tool for ZMK keyboards."
authors = [{name = "Kevin Hester", email = "kevinh@geeksville.com"}]
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.12,<3.15"
dependencies = [
    "keymap-drawer >= 0.22.1",
    "evdev >= 1.7; sys_platform == 'linux'",
    "pynput >= 1.7",
    "PyGObject >= 3.50.0",
    "pyyaml ~= 6.0",
    "bleak >= 0.22.0",
    "platformdirs >= 4.0",
]

[project.optional-dependencies]
dev = ["mypy", "pylint", "black", "types-pyyaml"]

[project.scripts]
zmk-buddy = 'zmk_buddy.__main__:main'

[project.urls]
homepage = "https://github.com/geeksville/zmk-buddy"
repository = "https://github.com/geeksville/zmk-buddy"
issues = "https://github.com/geeksville/zmk-buddy/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
requires-poetry = ">=2.0"

[tool.pylint.master]
extension-pkg-whitelist = ["pydantic", "gi"]

[tool.pylint.basic]
good-names = "x,y,w,h,r,f,k,v,p,m,c,a0"
max-line-length = 120

[tool.pylint."messages control"]
disable = [
    "too-few-public-methods",
    "line-too-long",
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-branches",
    "too-many-nested-blocks",
    "no-name-in-module",
    "import-outside-toplevel",
    "logging-fstring-interpolation",
    "broad-exception-caught",
    "invalid-name",
    "protected-access",
    "inconsistent-return-statements",
    "unused-argument",
    "wrong-import-order",
    "possibly-used-before-assignment",
]

[tool.mypy]
plugins = "pydantic.mypy"

[[tool.mypy.overrides]]
module = "gi.*"
ignore_missing_imports = true

[tool.black]
line-length = 120

[tool.isort]
src_paths = ["zmk_buddy"]
profile = "black"
line_length = 120

[dependency-groups]
dev = [
    "mypy (>=1.0)",
    "PyGObject-stubs (>=2.12.0)",
    "types-pynput (>=1.8.1.20250809,<2.0.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]
